Refactor Advanced acceptance tests and utils
