+++ airbyte-test-utils/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     implementation project(':airbyte-api')
     implementation project(':airbyte-workers')
 
+    implementation 'io.fabric8:kubernetes-client:5.12.2'
     implementation 'io.temporal:temporal-sdk:1.8.1'
